fix(crm): el cron de la caducidad llevaba cuatro días sin poder llamar a su función
crm.reactivar_bots se declaró con `p_dias smallint` y su cron quedó
registrado como `SELECT crm.reactivar_bots(3)`. El literal 3 es INTEGER, y
Postgres no convierte integer → smallint al resolver a qué función llamar:

    ERROR: function crm.reactivar_bots(integer) does not exist

Las cuatro corridas desde el 18 sep fallaron. La caducidad del silencio
estuvo muerta todo ese tiempo, con 4 contactos ya vencidos a los que el
bot nunca volvió. Nadie se enteró porque un cron que falla no avisa a
nadie: queda en cron.job_run_details y ahí se queda.

POR QUÉ LA PRUEBA NO LO VIO, QUE ES LO QUE IMPORTA

Llamaba `crm.reactivar_bots(3::smallint)`. Estaba escrita contra la
IMPLEMENTACIÓN y no contra la LLAMADA REAL: probaba que la función
funciona, nunca que alguien pudiera invocarla. Pasaba en verde mientras
producción llevaba días rota.

El arreglo de la prueba vale más que el de la función. Ahora hay una que
ejecuta el comando EXACTO registrado en cron.job, leyéndolo de la tabla:
si alguien cambia una firma y deja el agendado atrás, se cae ahí. Se
comprobó que caza el bug recreando la firma rota en local — una prueba
nueva que solo pasa sobre código arreglado no demuestra nada.

La función pasa a `int`, que es la convención de la casa
(cerrar_fantasmas(p_dias int) ya lo era) y lo que cualquiera escribe al
llamarla a mano.

319 pruebas en verde.

+++ supabase/tests/17_bot_por_lead.sql
@@ -232,7 +232,7 @@ VALUES ('b0700013-0000-0000-0000-000000000013',
         '11111111-1111-1111-1111-111111111111', now());
 
 SELECT is(
-  crm.reactivar_bots(3::smallint), 2,
+  crm.reactivar_bots(3), 2,
   'Caducan exactamente dos: el abandonado y el que solo miraron'
 );
